chore: relocated and consolidated DTOs for better modularity and maintainability
- Moved `OsvVulnerabilityDto`, `ProjectDto`, and `TlcConfig` to `TlcDto` and `OsvScannerDto` to reduce redundancy
- Updated tests to reflect changes in DTO structures
- Enhanced `TrivyAdapterTest` for improved exception handling and severity validation
- Added a TODO for `TlcAdapter` to adjust return type for libyears integration
